Load Microsoft truststore from Common CA database
